Use onPreventDefault for mousedown in pickers.
Firefox has a different behavior for elements that have a fixed
positioned element inside one that has tabindex.

Compared to other browsers clicking on the fixed positioned element does
not focus the element with tabindex, so preventing the mousedown event
is the only way to prevent the blur event from firing.